Homework 4/services/analysis/technical/requirements.txt
setuptools
aiohttp
asyncpg
pandas
numpy
ta-lib
python-dotenv 